Implement --no_cleanup option for test & serve commands.
Prevents deletion of temporary files created for/by Galaxy. Might be useful for debugging stuff (https://github.com/galaxyproject/planemo/issues/80#issuecomment-74892668) - probably would be a whole lot more useful for debugging if we also dropped a little shell script in that directory that allowed resuming Galaxy from that state.

+++ planemo/galaxy_config.py
@@ -200,7 +200,8 @@ def galaxy_config(ctx, tool_path, for_tests=False, **kwds):
 
         yield GalaxyConfig(galaxy_root, config_directory, env, test_data_dir)
     finally:
-        if created_config_directory:
+        cleanup = not kwds.get("no_cleanup", False)
+        if created_config_directory and cleanup:
             shutil.rmtree(config_directory)
